// src/apps/permissions/emuns/permission.enum.ts
export enum PermissionMessagesEnum {
  PERMISSION_NAME_INVALID = 'permissions.permission_name_invalid.error',
  key_INVALID = 'permissions.key_invalid.error',
  API_KEY_INVALID = 'permissions.api_key_invalid.error',
  GROUP_KEY_INVALID = 'permissions.group_key_invalid.error',
  PERMISSION_UUID_IS_INVALID = 'permissions.permission_uuid_invalid.error',
  NOT_FOUND = 'permissions.not_found.error',
  ALREADY_EXIST = 'permissions.already_exist.error',
  PERMISSIONS_INVALID = 'permissions.permissions_invalid.error',
  API_KEYS_REPEAT = 'permissions.api_keys_repeat.error',
  NAME_KEYS_REPEAT = 'permissions.name_keys_repeat.error',
  GROUP_NAME_INVALID = 'permissions.group_name_invalid.error',
  INTERNAL_INVALID = 'permissions.internal_invalid.error',
}